requirements.txt
flask
werkzeug
flask-restful
flask-httpauth
flask-jwt-extended
flask_sqlalchemy
validate_email
requests
passlib
psycopg2-binary
flask-cors
paho-mqtt
simplejson
pika
flask-bcrypt
resources
python-dateutil
flask-mqtt
flask-mail
rdp
cryptography
marshmallow==2.19.5
boto3
gunicorn
eventlet
flask_socketio
kombu
Shapely
py3DNS
